Gift

+++ team-utilisation-monitor/libs/repository/data-access/src/lib/data-access.repository.ts
@@ -27,14 +27,13 @@ export class DataAccessRepository {
         user_person.suspended=suspended;
         user_person.company_name=company;
         user_person.company_id=company_id;
-        
+
         user_person.positions=[];
-        
+
         const pos_obj=new PositionEntity();
         pos_obj.position=position;
 
         user_person.positions.push(pos_obj);
-
 
         return user_person;
 
@@ -182,11 +181,11 @@ export class DataAccessRepository {
                             }
                           }
                         }
-                      }  
+                      }
                     },include:{
                       positions:true
                     }
-                })      
+                })
 
                 //return new admin
 
@@ -498,29 +497,55 @@ export class DataAccessRepository {
 
           for(let i=0;i<projects.length;i++)
           {
-            const projectName=await (await this.getProject(projects[i].project_id)).project_name;
+            const projectName=(await this.getProject(projects[i].project_id)).project_name;
             await this.ResetAssignedHours(projectName);  //Reset the assigned hours for all teams on project
 
-            await this.prisma.team.delete(
+
+            await this.prisma.teamsOnProjects.deleteMany(
               {
                 where:
                 {
-                  id:id
+                  team_id:id,
+                  project_id:projects[i].project_id
                 }
               }
             )
 
             this.CalculateUtilizationVProject(projectName);
 
+
+            if(i==(projects.length-1))
+            {
+              await this.prisma.personOnTeams.deleteMany(
+                {
+                  where:
+                  {
+                    team_id:id
+                  }
+                }
+              )
+
+              await this.prisma.team.delete(
+                  {
+                    where:
+                    {
+                      id:id
+                    }
+                  }
+                )
+            }
+
           }
-        }
+
+
+
+          return "Team  Deleted"
+
+      }
       catch(e)
       {
-          if(e instanceof Prisma.PrismaClientKnownRequestError)
-          {
+        return "Team Deletion Failed";
 
-            return "Team Deletion Failed";
-          }
       }
     }
 
@@ -1159,8 +1184,8 @@ export class DataAccessRepository {
         });
 
         const people_arr=[];
-        
-        
+
+
         if(people)
         {
           for(let i=0;i<people.length;++i)
@@ -1178,10 +1203,10 @@ export class DataAccessRepository {
             person.approved=people[i].approved;
             person.company_name=((await this.getCompanyVID(people[i].company_id)).company_name);
             person.utilisation=people[i].utilisation;
-            
+
             //positions
             if(people[i].positions)
-            {   
+            {
               for(let k=0;k<people[i].positions.length;++k){
                 const pos_obj=new PositionEntity();
 
@@ -1198,7 +1223,7 @@ export class DataAccessRepository {
               for(let k=0;k<people[i].skills.length;++k)
               {
                 const skills_obj=new Skill();
-  
+
                 skills_obj.id=people[i].skills[k].skill_id;
                 skills_obj.skill=((await this.GetSkillVID(people[i].skills[k].skill_id)).skill);
 
@@ -1206,7 +1231,7 @@ export class DataAccessRepository {
               }
             }
             people_arr.push(person);
-          } 
+          }
         }
         else
             console.log("Object people returned null");
@@ -1221,7 +1246,7 @@ export class DataAccessRepository {
      * it gets. Returns Unknown if it's unable to find the position.
     */
     async getPositionVID(pos_id:number):Promise<string>{
-      
+
       const position=await this.prisma.position.findUnique({
         where:{
           id:pos_id
@@ -1234,7 +1259,7 @@ export class DataAccessRepository {
       else{
         return "Unknown";
       }
-      
+
     }
 
     /***
@@ -1242,9 +1267,9 @@ export class DataAccessRepository {
      * position doesn't exist.
     */
 
-    async getOnePersonVEmail(arg_email:string):Promise<UserPerson|string>
+    async getOnePersonVEmail(arg_email:string):Promise<UserPerson>
     {
-      
+
 
       try
       {
@@ -1255,39 +1280,28 @@ export class DataAccessRepository {
             include:{
                 positions:true,
                 company:true,
-                project:true,
+                project:true
             }
         })
 
-          let local_project:string;
-          let local_company:string;
-          let title:string;
+        let ReturnPerson=new UserPerson();
+        ReturnPerson.id=person.id;
+        ReturnPerson.name=person.name;
+        ReturnPerson.surname=person.surname;
+        ReturnPerson.email=person.email;
+        ReturnPerson.approved=person.approved;
+        ReturnPerson.company_name=person.company.company_name;
+        ReturnPerson.role=person.role;
+        ReturnPerson.utilisation=person.utilisation
 
+        return ReturnPerson;
 
-          if(person.project==null)
-              local_project=null;
-          else
-              local_project=person.project.project_name
-
-          if(person.company==null)
-              local_company=null;
-          else
-              local_company=person.company.company_name;
-
-          if(person.positions!=null)
-            title=(await this.getPositionVID(person.positions[person.positions.length-1].id)); //latest position.
-
-          const return_user= await this.returnObject(person.id,person.name,person.surname,person.email,person.suspended,person.role,local_company,title,person.company_id);
-
-          return_user.utilisation=person.utilisation;
-          return_user.approved=person.approved;
-          return return_user;
       }
       catch(e)
       {
         if(e instanceof Prisma.PrismaClientKnownRequestError)
         {
-          return "Email not found"
+          return null
         }
       }
     }
@@ -1983,13 +1997,13 @@ export class DataAccessRepository {
 
         const obj=new UserPerson()
         obj.positions=[];
-        
+
 
         obj.name=person.name;
         obj.surname=person.surname;
         obj.email=person.email;
         obj.utilisation=person.utilisation;
-        
+
         if(person.positions.length>0)
         {
           for(let k=0;k<person.positions.length;++k)
@@ -1999,13 +2013,13 @@ export class DataAccessRepository {
               const pos_object=new PositionEntity();
               pos_object.team_name=person.positions[k].team_name;
               pos_object.position=(await this.getPositionVID(person.positions[k].position_id));
-              
+
               obj.positions.push(pos_object);
             }
           }
 
           //if the user has positions, but not in this team.
-          if(obj.positions.length==0) 
+          if(obj.positions.length==0)
           {
             const pos_object=new PositionEntity();
             pos_object.team_name=teamName;
@@ -2236,14 +2250,14 @@ export class DataAccessRepository {
           })
 
           return new MessageObject("SUCCESS"); //ErrorString is NONE
-          
+
         } catch (err) {
           if(err instanceof Prisma.PrismaClientKnownRequestError){
             return new MessageObject("Unable to create a position",ErrorStrings.PRISMA_CREATE_FAIL);
           }
         }
 
-        
+
       }
     }
 
@@ -2255,7 +2269,7 @@ export class DataAccessRepository {
     */
 
     async assignPositionToUser(email:string, position_name:string,teamName:string):Promise<MessageObject>
-    { 
+    {
       const p_id=await this.getPersonIDVEmail(email);
       const pos_id=await this.getPositionIDVName(position_name);
 
@@ -2295,7 +2309,7 @@ export class DataAccessRepository {
 
       const positions=await this.prisma.position.findMany();
       let return_arr:PositionEntity[]=[];
-      
+
 
       if(positions!=null)
       {
@@ -2520,7 +2534,7 @@ export class DataAccessRepository {
         obj.error_string=ErrorStrings.EMAIL_DOESNT_EXISTS;
         return_arr.push(obj);
       }
-      
+
       return return_arr;
     }
 
@@ -4260,5 +4274,6 @@ export class DataAccessRepository {
         }
       }
     }
+
 
 }
